pyproject.toml
[project]
name = "senter"
version = "3.0.0"
description = "Unified Personal AI with Cognitive Coupling"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "httpx>=0.27.0",
    "rich>=13.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
gguf = ["llama-cpp-python>=0.3.0"]
openai = ["openai>=1.0"]
ollama = ["ollama>=0.4.0"]
tui = ["textual>=0.89.0"]
voice = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
]
gaze = [
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
]
multimodal = [
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
]
all = [
    "llama-cpp-python>=0.3.0",
    "openai>=1.0",
    "ollama>=0.4.0",
    "textual>=0.89.0",
    "openai-whisper>=20231117",
    "sounddevice>=0.4.6",
    "mediapipe>=0.10.0",
    "opencv-python>=4.8.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[project.scripts]
senter = "senter:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
